Move EmailAddress into models/user/
