+++ models/user_mail.go
@@ -1,5 +1,4 @@
-// Copyright 2016 The Gogs Authors. All rights reserved.
-// Copyright 2020 The Gitea Authors. All rights reserved.
+// Copyright 2021 The Gitea Authors. All rights reserved.
 // Use of this source code is governed by a MIT-style
 // license that can be found in the LICENSE file.
 
@@ -7,179 +6,28 @@ package models
 
 import (
 	"fmt"
-	"net/mail"
 	"strings"
 
 	"code.gitea.io/gitea/models/db"
-	"code.gitea.io/gitea/modules/log"
-	"code.gitea.io/gitea/modules/setting"
+	user_model "code.gitea.io/gitea/models/user"
 	"code.gitea.io/gitea/modules/util"
-
 	"xorm.io/builder"
 )
 
-// EmailAddress is the list of all email addresses of a user. It also contains the
-// primary email address which is saved in user table.
-type EmailAddress struct {
-	ID          int64  `xorm:"pk autoincr"`
-	UID         int64  `xorm:"INDEX NOT NULL"`
-	Email       string `xorm:"UNIQUE NOT NULL"`
-	LowerEmail  string `xorm:"UNIQUE NOT NULL"`
-	IsActivated bool
-	IsPrimary   bool `xorm:"DEFAULT(false) NOT NULL"`
-}
-
-func init() {
-	db.RegisterModel(new(EmailAddress))
-}
-
-// BeforeInsert will be invoked by XORM before inserting a record
-func (email *EmailAddress) BeforeInsert() {
-	if email.LowerEmail == "" {
-		email.LowerEmail = strings.ToLower(email.Email)
-	}
-}
-
-// ValidateEmail check if email is a allowed address
-func ValidateEmail(email string) error {
-	if len(email) == 0 {
-		return nil
-	}
-
-	if _, err := mail.ParseAddress(email); err != nil {
-		return ErrEmailInvalid{email}
-	}
-
-	// TODO: add an email allow/block list
-
-	return nil
-}
-
-// GetEmailAddresses returns all email addresses belongs to given user.
-func GetEmailAddresses(uid int64) ([]*EmailAddress, error) {
-	emails := make([]*EmailAddress, 0, 5)
-	if err := db.GetEngine(db.DefaultContext).
-		Where("uid=?", uid).
-		Asc("id").
-		Find(&emails); err != nil {
-		return nil, err
-	}
-	return emails, nil
-}
-
-// GetEmailAddressByID gets a user's email address by ID
-func GetEmailAddressByID(uid, id int64) (*EmailAddress, error) {
-	// User ID is required for security reasons
-	email := &EmailAddress{UID: uid}
-	if has, err := db.GetEngine(db.DefaultContext).ID(id).Get(email); err != nil {
-		return nil, err
-	} else if !has {
-		return nil, nil
-	}
-	return email, nil
-}
-
-// isEmailActive check if email is activated with a different emailID
-func isEmailActive(e db.Engine, email string, excludeEmailID int64) (bool, error) {
-	if len(email) == 0 {
-		return true, nil
-	}
-
-	// Can't filter by boolean field unless it's explicit
-	cond := builder.NewCond()
-	cond = cond.And(builder.Eq{"lower_email": strings.ToLower(email)}, builder.Neq{"id": excludeEmailID})
-	if setting.Service.RegisterEmailConfirm {
-		// Inactive (unvalidated) addresses don't count as active if email validation is required
-		cond = cond.And(builder.Eq{"is_activated": true})
-	}
-
-	var em EmailAddress
-	if has, err := e.Where(cond).Get(&em); has || err != nil {
-		if has {
-			log.Info("isEmailActive(%q, %d) found duplicate in email ID %d", email, excludeEmailID, em.ID)
-		}
-		return has, err
-	}
-
-	return false, nil
-}
-
-func isEmailUsed(e db.Engine, email string) (bool, error) {
-	if len(email) == 0 {
-		return true, nil
-	}
-
-	return e.Where("lower_email=?", strings.ToLower(email)).Get(&EmailAddress{})
-}
-
-// IsEmailUsed returns true if the email has been used.
-func IsEmailUsed(email string) (bool, error) {
-	return isEmailUsed(db.GetEngine(db.DefaultContext), email)
-}
-
-func addEmailAddress(e db.Engine, email *EmailAddress) error {
-	email.Email = strings.TrimSpace(email.Email)
-	used, err := isEmailUsed(e, email.Email)
-	if err != nil {
-		return err
-	} else if used {
-		return ErrEmailAlreadyUsed{email.Email}
-	}
-
-	if err = ValidateEmail(email.Email); err != nil {
-		return err
-	}
-
-	_, err = e.Insert(email)
-	return err
-}
-
-// AddEmailAddress adds an email address to given user.
-func AddEmailAddress(email *EmailAddress) error {
-	return addEmailAddress(db.GetEngine(db.DefaultContext), email)
-}
-
-// AddEmailAddresses adds an email address to given user.
-func AddEmailAddresses(emails []*EmailAddress) error {
-	if len(emails) == 0 {
-		return nil
-	}
-
-	// Check if any of them has been used
-	for i := range emails {
-		emails[i].Email = strings.TrimSpace(emails[i].Email)
-		used, err := IsEmailUsed(emails[i].Email)
-		if err != nil {
-			return err
-		} else if used {
-			return ErrEmailAlreadyUsed{emails[i].Email}
-		}
-		if err = ValidateEmail(emails[i].Email); err != nil {
-			return err
-		}
-	}
-
-	if _, err := db.GetEngine(db.DefaultContext).Insert(emails); err != nil {
-		return fmt.Errorf("Insert: %v", err)
-	}
-
-	return nil
-}
-
-// Activate activates the email address to given user.
-func (email *EmailAddress) Activate() error {
+// ActivateEmail activates the email address to given user.
+func ActivateEmail(email *user_model.EmailAddress) error {
 	sess := db.NewSession(db.DefaultContext)
 	defer sess.Close()
 	if err := sess.Begin(); err != nil {
 		return err
 	}
-	if err := email.updateActivation(sess, true); err != nil {
+	if err := updateActivation(sess, email, true); err != nil {
 		return err
 	}
 	return sess.Commit()
 }
 
-func (email *EmailAddress) updateActivation(e db.Engine, activate bool) error {
+func updateActivation(e db.Engine, email *user_model.EmailAddress, activate bool) error {
 	user, err := getUserByID(e, email.UID)
 	if err != nil {
 		return err
@@ -194,58 +42,17 @@ func (email *EmailAddress) updateActivation(e db.Engine, activate bool) error {
 	return updateUserCols(e, user, "rands")
 }
 
-// DeleteEmailAddress deletes an email address of given user.
-func DeleteEmailAddress(email *EmailAddress) (err error) {
-	if email.IsPrimary {
-		return ErrPrimaryEmailCannotDelete{Email: email.Email}
-	}
-
-	var deleted int64
-	// ask to check UID
-	address := EmailAddress{
-		UID: email.UID,
-	}
-	if email.ID > 0 {
-		deleted, err = db.GetEngine(db.DefaultContext).ID(email.ID).Delete(&address)
-	} else {
-		if email.Email != "" && email.LowerEmail == "" {
-			email.LowerEmail = strings.ToLower(email.Email)
-		}
-		deleted, err = db.GetEngine(db.DefaultContext).
-			Where("lower_email=?", email.LowerEmail).
-			Delete(&address)
-	}
-
-	if err != nil {
-		return err
-	} else if deleted != 1 {
-		return ErrEmailAddressNotExist{Email: email.Email}
-	}
-	return nil
-}
-
-// DeleteEmailAddresses deletes multiple email addresses
-func DeleteEmailAddresses(emails []*EmailAddress) (err error) {
-	for i := range emails {
-		if err = DeleteEmailAddress(emails[i]); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
 // MakeEmailPrimary sets primary email address of given user.
-func MakeEmailPrimary(email *EmailAddress) error {
+func MakeEmailPrimary(email *user_model.EmailAddress) error {
 	has, err := db.GetEngine(db.DefaultContext).Get(email)
 	if err != nil {
 		return err
 	} else if !has {
-		return ErrEmailAddressNotExist{Email: email.Email}
+		return user_model.ErrEmailAddressNotExist{Email: email.Email}
 	}
 
 	if !email.IsActivated {
-		return ErrEmailNotActivated
+		return user_model.ErrEmailNotActivated
 	}
 
 	user := &User{}
@@ -269,7 +76,7 @@ func MakeEmailPrimary(email *EmailAddress) error {
 	}
 
 	// 2. Update old primary email
-	if _, err = sess.Where("uid=? AND is_primary=?", email.UID, true).Cols("is_primary").Update(&EmailAddress{
+	if _, err = sess.Where("uid=? AND is_primary=?", email.UID, true).Cols("is_primary").Update(&user_model.EmailAddress{
 		IsPrimary: false,
 	}); err != nil {
 		return err
@@ -347,7 +154,7 @@ func SearchEmails(opts *SearchEmailOptions) ([]*SearchEmailResult, int64, error)
 	}
 
 	count, err := db.GetEngine(db.DefaultContext).Join("INNER", "`user`", "`user`.ID = email_address.uid").
-		Where(cond).Count(new(EmailAddress))
+		Where(cond).Count(new(user_model.EmailAddress))
 	if err != nil {
 		return nil, 0, fmt.Errorf("Count: %v", err)
 	}
@@ -374,15 +181,16 @@ func SearchEmails(opts *SearchEmailOptions) ([]*SearchEmailResult, int64, error)
 // ActivateUserEmail will change the activated state of an email address,
 // either primary or secondary (all in the email_address table)
 func ActivateUserEmail(userID int64, email string, activate bool) (err error) {
-	sess := db.NewSession(db.DefaultContext)
-	defer sess.Close()
-	if err = sess.Begin(); err != nil {
+	ctx, committer, err := db.TxContext()
+	if err != nil {
 		return err
 	}
+	defer committer.Close()
+	sess := db.GetEngine(ctx)
 
 	// Activate/deactivate a user's secondary email address
 	// First check if there's another user active with the same address
-	addr := EmailAddress{UID: userID, LowerEmail: strings.ToLower(email)}
+	addr := user_model.EmailAddress{UID: userID, LowerEmail: strings.ToLower(email)}
 	if has, err := sess.Get(&addr); err != nil {
 		return err
 	} else if !has {
@@ -393,13 +201,13 @@ func ActivateUserEmail(userID int64, email string, activate bool) (err error) {
 		return nil
 	}
 	if activate {
-		if used, err := isEmailActive(sess, email, addr.ID); err != nil {
+		if used, err := user_model.IsEmailActive(ctx, email, addr.ID); err != nil {
 			return fmt.Errorf("unable to check isEmailActive() for %s: %v", email, err)
 		} else if used {
-			return ErrEmailAlreadyUsed{Email: email}
+			return user_model.ErrEmailAlreadyUsed{Email: email}
 		}
 	}
-	if err = addr.updateActivation(sess, activate); err != nil {
+	if err = updateActivation(sess, &addr, activate); err != nil {
 		return fmt.Errorf("unable to updateActivation() for %d:%s: %w", addr.ID, addr.Email, err)
 	}
 
@@ -423,5 +231,5 @@ func ActivateUserEmail(userID int64, email string, activate bool) (err error) {
 		}
 	}
 
-	return sess.Commit()
+	return committer.Commit()
 }
